+++ src/simmate/calculators/vasp/database/dynamics.py
@@ -6,3 +6,16 @@ MITDynamicsRun, MITDynamicsIonicStep = DynamicsRun.create_subclasses(
     "MIT",
     module=__name__,
 )
+
+MatprojDynamicsRun, MatprojDynamicsIonicStep = DynamicsRun.create_subclasses(
+    "Matproj",
+    module=__name__,
+)
+
+(
+    MatVirtualLabNPTDynamicsRun,
+    MatVirtualLabNPTDynamicsIonicStep,
+) = DynamicsRun.create_subclasses(
+    "MatVirtualLabNPT",
+    module=__name__,
+)