+++ CHANGELOG.md
@@ -22,6 +22,7 @@ There is one key exception to the rules above -- and that is with `MAJOR`=0 rele
 **Enhancements**
 - NEB workflows now accept parameters for changing supercell size and number of images used
 - add HSE workflows for static energy, relaxation, and DOS/BS electronic states
+- add NPT and MatProj molecular dynamics workflows
 
 **Refactors**
 - None